another test, but still bad

+++ gang_decoder_impl.h
@@ -19,13 +19,13 @@ struct gang_decoder {
 	enum AVPixelFormat pix_fmt;
 	// audio
 
-	int sample_rate;
-	int channels;
-	int bytesPerSample;
+	uint32_t sample_rate;
+	uint8_t channels;
+	int8_t bytesPerSample;
 
-	 uint8_t *video_dst_data[4];
-	 int      video_dst_linesize[4];
-	 int video_dst_bufsize;
+	uint8_t *video_dst_data[4];
+	int video_dst_linesize[4];
+	int video_dst_bufsize;
 
 	AVFormatContext *i_fmt_ctx;
 
@@ -59,7 +59,10 @@ int start_gang_decode(struct gang_decoder* decoder);
 // output: *data, need allocate and assign to **data.(do not free it)
 //         size
 // return: -1->EOF, 0->error, 1->video, 2->audio
-int gang_decode_next_frame(struct gang_decoder* decoder, void **data, int *size);
+int gang_decode_next_frame(
+		struct gang_decoder* decoder,
+		void **data,
+		int *size);
 
 // disconnect from remote stream and free AVCodecContext...
 void stop_gang_decode(struct gang_decoder* decoder);
